+++ tests/supervisors/test_watchdogreload.py
@@ -3,25 +3,18 @@ import signal
 import time
 from pathlib import Path
 
-import pytest
-
 from uvicorn.config import Config
-from uvicorn.supervisors.statreload import StatReload
+from uvicorn.main import Server
+from uvicorn.supervisors.watchdogreload import WatchdogReload
 
 
 def run(sockets):
     pass
 
 
-def test_statreload():
-    """
-    A basic sanity check.
-
-    Simply run the reloader against a no-op server, and signal for it to
-    quit immediately.
-    """
-    config = Config(app=None, reload=True)
-    reloader = StatReload(config, target=run, sockets=[])
+def test_statreload(certfile_and_keyfile):
+    config = Config(app=None)
+    reloader = WatchdogReload(config, target=run, sockets=[])
     reloader.signal_handler(sig=signal.SIGINT, frame=None)
     reloader.run()
 
@@ -34,13 +27,13 @@ def test_should_reload(tmpdir):
     os.chdir(str(tmpdir))
     try:
         config = Config(app=None, reload=True)
-        reloader = StatReload(config, target=run, sockets=[])
+        reloader = WatchdogReload(config, target=run, sockets=[])
         reloader.signal_handler(sig=signal.SIGINT, frame=None)
         reloader.startup()
 
         assert not reloader.should_restart()
-        time.sleep(0.1)
         update_file.touch()
+        time.sleep(0.1)
         assert reloader.should_restart()
 
         reloader.restart()