+++ tests/supervisors/__init__.py
@@ -1,0 +1,5 @@
+WATCHED_FILES = (
+    "example.py",
+    "example.html",
+    "example.graphql",
+)