+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Use `watchdog`, if installed, for watching code changes.
+* Use `watchgod`, if installed, for watching code changes.
 
 ## 0.11.3