+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Use `watchgod`, if installed, for watching code changes.
+* Reload application when any files in watched directories change, not just `.py` files.
 
 ## 0.11.3